requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.4
django-crispy-forms==2.1
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
django-parler==2.3
flower==1.2.0
celery==5.2.7
stripe==4.0.2
WeasyPrint==56.1
redis==4.3.4
django-rosetta==0.9.8
django-localflavor==3.1
